backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
boto3==1.29.3
python-dotenv==1.0.0
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
psycopg2-binary==2.9.9
gunicorn==20.1.0 